feat: Add an internal function to convert file URL
This function behave the same as NodeJS `url.fileURLToPath` function, but that makes this function differ is that this function supports relative paths being passed in the argument (e.g., "file:./foo/bar"). If the given URL is have non-file protocols or contains invalid path structures, an URIError will be thrown.

This function will be used as part of `ls*` functions to support file URL paths. In addition to this change, I've also added the documentation for the function.