refactor(ls): Fix absolute path resolver on Windows
The path passed to `fs.readdir` function can't be a Windows path, instead it must a node path (i.e., similar to POSIX path). That's why we need to resolve the relative path to absolute with format of POSIX path, so the path can be used to the `fs.readdir` function without getting an internal error.