fix: Fix unrecognized type when using the string value
Previously, if using the string value of `lsTypes` enum (e.g., 'LS_F') the `ls` function can't recognizes it correctly and ignore it, thus causing the returned value is the undesired entries result. This change is to address the issue and now the `ls` function correctly recognized the string value of `lsTypes` enum, improving flexibility to the function.

As of this change, you can now call the `ls` function with less code and more flexible:

  const files = await ls('./foo', {}, 'LS_F');