refactor: Improve the type argument checker
In the last change we have configured the `ls` function to accepts the string value representing the `lsTypes` property names as value of `type` argument. This change brings the type argument checker to more strict and prevent from unintended values. In addition to this change, a new helper function called `checkType` has been defined to checks and verifies the type argument and compare it with the valid types.