refactor: Reorganize and improve types declaration
This change reorganizes and refactors the types declaration within the `types` module for better maintainability and future development. It includes restructuring the type definitions for specific API parameters.

Additionally, the `types` module is now re-exported in the main package entry (`index.ts`), and the `declaration` option is enabled in the 'tsconfig.production.json' file.

These changes enable the LSFND package to be correctly imported and used in TypeScript projects with the `module` set to `node16`, `nodenext`, `es5`, `esnext`, and others, in addition to `commonjs`, which was supported in previous release.