feat(api): Add new several API options
Here is the list new options that just being added in this change:
  - absolute?: { boolean } - Whether to return absolute paths for all entries.
  - basename? { boolean } - Whether to makes the returned entries only have their file and/or directories names.
  - rootDir?: { StringPath } - A string path represents the root directory to resolve the relative paths.

They all were have their own priority tier, in the list above their priority are sorted from the highest (top) to the lowest (bottom). So, if you specified the `rootDir` but you also enabled the `absolute` option, then -- as a result, -- the `absolute` will be used instead (it is because the `absolute` option have the highest priority) and the `rootDir` value will be ignored.

Please note, that the `basename` option are implicitly includes any directory names while listing a directory. If you want include the filenames only, then combining it with the `lsTypes.LS_F` is a good way if you are using `ls` function, or combine this option with `lsFiles` function for more flexible.