fix(encoding): Resolve encoding usage in the `ls` function
Previously, the encoding option are fully malfunctioned and can't be passed with other encodings except 'utf8', 'binary', and 'ascii' (in some cases). If we specify with 'hex' encoding, the function will throws an error (which is due to internal error) complaining an invalid path on the specific entry. Now it has been resolved and work as expected, thankfully to `encodeTo` internal function to encode the entries and vice versa, this also includes improving the resolution logic of `encoding` option to more strict and robust.