fix(build): Resolve bad shell invocation shell
Before, during the building procedure, there was a problem with the invocation shell in the child process that used `child_process.spawn`. In Windows, this resulted in the process issuing an ENOENT error since the command was not recognized. However, with this modification, we resolved the issue and it now functions correctly in Windows by activating the `shell` option within the options argument (3rd argument) of the `spawn` function.

Note:
Tested in PowerShell and MSYS2 MinTTY. This change are exclusively to Windows users and developers.