feat: Add new function to resolve file URL path
Added a new function to resolve file URL path to a file path. The given path should be a valid file URL following the WHATWG URL Standard.

If the provided URL is 'file://' or 'file:///', it is replaced with the root directory path (in the current drive for Windows systems). Otherwise, the URL is parsed using the `fileURLToPath` function.