fix: Fix error while stating read-protected entry
Fixed the error while attempting to stat an entry that are read-protected due to a permission error (EPERM) or access error (EACCES).

If throws an error during stating, then it attempts to open the entry using `fs.opendir` , if throw an error with code ENOTDIR we can coclude that the entry is a regular file, otherwise, it is a directory. We also not forget to close the `Dir` instance to avoid unexpected behavior. In addition, we also sort the result entries exact before it being returned.